Add move useRtl hook to locale folder
