+++ packages/common/src/styles/AppGlobalStyles.test.tsx
@@ -7,7 +7,7 @@ import { AppGlobalStyles } from './AppGlobalStyles';
 describe('AppGlobalStyles', () => {
   it('Sets the direction of the body to be rtl when a rtl language is the current locale', () => {
     act(() => {
-      LocalStorage.setItem('/localisation/locale', 'fr');
+      LocalStorage.setItem('/localisation/locale', 'ab');
     });
 
     const { baseElement } = render(<AppGlobalStyles />);
